+++ mednafen/psx/input/justifier.cpp
@@ -169,7 +169,7 @@ int InputDevice_Justifier::StateAction(StateMem* sm, int load, int data_only, co
 
    if(load)
    {
-      if((transmit_pos + transmit_count) > sizeof(transmit_buffer))
+      if(((uint64_t)transmit_pos + transmit_count) > sizeof(transmit_buffer))
       {
          transmit_pos = 0;
          transmit_count = 0;
